feat: Add DeepSeek attention, MLA v2, TurboQuant page attention, and grouped matmul v3 operations
Registers new kernels (deepseek_attn, grouped_matmulv3,
multi_latent_ragged_page_attention_v2, ragged_page_attention_v2/v3_turboquant)
with configs, module exports, and tests. Extends GroupedMatmul with version-based
dispatch and rhs_scale/rhs_bias support for v3. Switches GDR Pallas precision to
DEFAULT and fixes XLA state dtype promotion. Bumps version to 0.0.75.

+++ ejkernel/kernels/_xla/gated_delta_rule/_xla_impl_fwd.py
@@ -373,7 +373,7 @@ def _chunk_gdr_fwd_core(
         state_decayed = state * g_end_exp_i[:, :, None, None]
         k_scaled = k_i * g_diff_exp_i[:, :, :, None]
         state_update = jnp.einsum("bhik,bhiv->bhkv", k_scaled, v_new, precision=_MATMUL_PRECISION)
-        new_state = state_decayed + state_update
+        new_state = (state_decayed + state_update).astype(state.dtype)
 
         return new_state, core_out.astype(input_dtype)
